requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
resend~=0.6.0
waitress~=3.0.2
Werkzeug~=3.1.3
bleach==6.0.0